rewrite func

+++ samoilenko.alena/I4/columns_without_null_elements.cpp
@@ -1,7 +1,6 @@
 #include "columns_without_null_elements.h"
-#include <cstddef>
 
-int countColumnsWithoutNull(const int* matrix, size_t row, size_t col)
+size_t countColumnsWithoutNull(const int* matrix, size_t row, size_t col)
 {
   int count = 0;
   for (size_t i = 0; i < col; i++)
